third_party/WebKit/LayoutTests/http/tests/htmlimports/import-preload.html

<!DOCTYPE html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
function testPreload(url, doc)
{
    if (internals.isPreloadedBy(url, doc)) {
        console.log("PASS: " + url + " is preloaded.");
    } else {
        console.log("FAIL: " + url + " should be preloaded");
    }
}
</script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
<script>
testPreload('resources/preload.html', document);
</script>
<link id=target rel=import href="resources/preload.html">
</head>
