third_party/WebKit/LayoutTests/http/tests/htmlimports/import-preload-expected.txt

CONSOLE MESSAGE: line 9: PASS: resources/preload.html is preloaded.
